backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Douglas Phelipe",email = "douglasphelipedpss@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10, <4"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "flask-jwt-extended (>=4.7.1,<5.0.0)",
    "flask-bcrypt",
    "flask-marshmallow",
    "marshmallow-sqlalchemy",
    "apispec-webframeworks",
    "flask-apispec[swagger]",
]

[tool.poetry]
packages = [{include = "backend", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
